shop/serializers.py
from rest_framework import serializers
from .models import Furniture, Image

class ImageSerializer(serializers.ModelSerializer):
    class Meta:
        model = Image
        fields = '__all__'

class FurnitureSerializer(serializers.ModelSerializer):
    images = ImageSerializer(many=True)

    class Meta:
        model = Furniture
        fields = '__all__'

